chore: move user info above tabs
