+++ apps/web/src/components/InboxTabs/Seen.tsx
@@ -18,7 +18,7 @@ export const Seen = () => {
   return (
     <section className='mb-8 flex flex-col items-center space-y-12'>
       {/* Top */}
-      <div className='border-secondary-100 bg-secondary-200 w-full overflow-hidden rounded-2xl border-2 md:w-[500px]'>
+      <div className='border-secondary-100 bg-secondary-200 w-full overflow-hidden rounded-2xl border-2'>
         <div className='border-secondary-100 flex items-center justify-between border-b-2 bg-[#171819] py-1'>
           <div className='relative mx-auto h-[40px] w-[120px]'>
             <Image src='/assets/logo.svg' layout='fill' objectFit='contain' />
