+++ apps/web/src/components/InboxTabs/Sent.tsx
@@ -9,7 +9,7 @@ export const Sent = () => {
   return (
     <section className='flex flex-col items-center'>
       {/* Top */}
-      <div className='border-secondary-100 bg-secondary-200 w-full overflow-hidden rounded-2xl border-2 md:w-[500px]'>
+      <div className='border-secondary-100 bg-secondary-200 w-full overflow-hidden rounded-2xl border-2'>
         <div className='border-secondary-100 flex items-center justify-between border-b-2 bg-[#171819] px-7 py-2'>
           <p className='font-medium capitalize text-gray-100'>
             <span className='font-light text-gray-400'>To&#58;</span> Johndoe
